Features/functions implementation

+++ src/Quant.Infra.Net/SourceData/Model/Ohlcv.cs
@@ -14,7 +14,7 @@ namespace Quant.Infra.Net.SourceData.Model
 
         public decimal Close { get; set; }
 
-        public long Volume { get; set; }
+        public decimal Volume { get; set; }
 
         public decimal AdjustedClose { get; set; }
     }
